new feature to manually chose locale.
right now the region was set to default value of "US", and basically for changing the locale we have to change it manually to whatever region we wanted. with this commit i have added an extra functionality for choosing the region on runtime. just you have to define region whilst you define other parameters like AssociateTag and awskey, like.

amazon = bottlenose.Amazon("AWSAccessKeyId", "AWSSecretAccessKey", "AssociateTag", "IN")